+scan sequence

+++ jfixby-cmns-api/jfixby-commons-api/com/jfixby/cmns/api/collections/CollectionScanner.java
@@ -3,6 +3,6 @@ package com.jfixby.cmns.api.collections;
 
 public interface CollectionScanner<T> {
 
-	public void scanElement (T element, int index);
+	public void scanElement (T element, long index);
 
 }
